feat(feedback): create generic not-found indicator (#283)
* feat(message): Lagt til bilde prop i MessageIndicator

feat(message): Lagt til  NoEvetntsIndicator, NoPostsIndicator, NoNewsIndicator og NoPersonsIndicator i MessageIndicator. Oppdatert til funksjoner fra klasser.

* fix import

* chore(dev): update from dev

* feat(notfound): update not found indicator

* Update NotFoundIndicator.tsx
